+++ src/main/java/com/beonse2/domain/point/mapper/PointMapper.java
@@ -1,10 +1,9 @@
-package com.beonse2.domain.point.repository;
+package com.beonse2.domain.point.mapper;
 
 import com.beonse2.domain.point.vo.PointVO;
 import org.apache.ibatis.annotations.Mapper;
 
 import java.util.List;
-import java.util.Optional;
 
 @Mapper
 public interface PointMapper {
