+++ src/main/resources/mapper/PointMapper.xml
@@ -2,7 +2,7 @@
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.beonse2.domain.point.repository.PointMapper">
+<mapper namespace="com.beonse2.domain.point.mapper.PointMapper">
 
     <insert id="save" parameterType="com.beonse2.domain.point.vo.PointVO">
         INSERT INTO point (pid, member_mid, points, payment_price, card_name, card_number, payment_date)
